requirements.txt
Pillow
torch
torchvision
openai-clip
transformers
tqdm